+++ drivers/web-worker-driver/src/commonMain/kotlin/app/cash/sqldelight/driver/worker/WebWorkerDriver.kt
@@ -21,14 +21,16 @@ import app.cash.sqldelight.driver.worker.expected.checkWorkerResults
  * This driver is dialect-agnostic and is instead dependent on the Worker script's implementation
  * to handle queries and send results back from the Worker.
  *
- * @property worker The Worker running a SQL implementation that this driver communicates with.
  * @see [createDefaultWebWorkerDriver]
  */
-class WebWorkerDriver(private val worker: Worker) : SqlDriver {
+class WebWorkerDriver internal constructor(
+  private val wrapper: WorkerWrapper,
+) : SqlDriver {
+  constructor(worker: Worker) : this(WorkerWrapper(worker))
+
   private val listeners = mutableMapOf<String, MutableSet<Query.Listener>>()
   private var messageCounter = 0
   private var transaction: Transacter.Transaction? = null
-  private val wrapper = WorkerWrapper(worker)
 
   override fun <R> executeQuery(
     identifier: Int?,
@@ -93,13 +95,13 @@ class WebWorkerDriver(private val worker: Worker) : SqlDriver {
 
   override fun newTransaction(): QueryResult<Transacter.Transaction> = QueryResult.AsyncValue {
     val enclosing = transaction
-    val transaction = Transaction(enclosing)
-    this.transaction = transaction
     if (enclosing == null) {
       wrapper.sendMessage(WorkerActions.beginTransaction)
     }
 
-    return@AsyncValue transaction
+    return@AsyncValue Transaction(enclosing).also {
+      transaction = it
+    }
   }
 
   override fun currentTransaction(): Transacter.Transaction? = transaction
@@ -108,14 +110,17 @@ class WebWorkerDriver(private val worker: Worker) : SqlDriver {
     override val enclosingTransaction: Transacter.Transaction?,
   ) : Transacter.Transaction() {
     override fun endTransaction(successful: Boolean): QueryResult<Unit> = QueryResult.AsyncValue {
-      if (enclosingTransaction == null) {
-        if (successful) {
-          wrapper.sendMessage(WorkerActions.endTransaction)
-        } else {
-          wrapper.sendMessage(WorkerActions.rollbackTransaction)
+      try {
+        if (enclosingTransaction == null) {
+          if (successful) {
+            wrapper.sendMessage(WorkerActions.endTransaction)
+          } else {
+            wrapper.sendMessage(WorkerActions.rollbackTransaction)
+          }
         }
+      } finally {
+        transaction = enclosingTransaction
       }
-      transaction = enclosingTransaction
     }
   }