+++ drivers/web-worker-driver/src/commonTest/kotlin/app/cash/sqldelight/drivers/worker/CreateControllableWebWorkerDriver.kt
@@ -1,0 +1,5 @@
+package app.cash.sqldelight.drivers.worker
+
+import app.cash.sqldelight.driver.worker.SqliteWasmWebWorkerDriver
+
+expect fun createControllableWebWorkerDriver(): SqliteWasmWebWorkerDriver